Original CTO.org
This is the original CTO.org code. The django.wsgi and wsgi_app.py files
are necessary for Apache to serve the site correctly.
The /django_cardiologytrials/ is the actual django directory. /www/
lists the static files also required for the website to be served
appropriately (images, styles (including those for django admin) etc.)